requirements.txt
Django==1.7
dj-database-url==0.2.2
django-nose==1.3
psycopg2==2.5.2
uWSGI==2.0
wsgiref==0.1.2
requests==2.2.1
djorm-ext-core==0.7
djorm-ext-expressions==0.6
djorm-ext-pgarray==0.10
boto==2.29.1
python-mimeparse==0.1.4
python-dateutil==2.2
django-tastypie==0.12.0
